Fix the frontend unit tests
It turns out that while the karma tests ran, it didn't actually run
the unit tests. This is because we had angular-scenario.js in our
public interface that was getting included in with the unit tests.
angular-scenario.js conflicts with angular-mocks.js, which ended
up sepressing the unit tests. This removes angular-scenario.js from
the public interface, which gets the unit tests to run again.