add exporting a graph subset to a new project.
Note this fixes the nstep subgraph, but it does so in a very
inefficient manner. We are currently walking N steps out from
each initial node, which may result in touching the same nodes
multiple times over. It would be more efficient to maintain a
queue of some sort to cut down on the number of visits we do
to each node.